+++ src/solver/CMakeLists.txt
@@ -15,6 +15,7 @@ add_subdirectory(infeasible-problem-analysis)
 add_subdirectory(lps)
 add_subdirectory(libModelObject)
 add_subdirectory(misc)
+add_subdirectory(modelConverter)
 add_subdirectory(modelParser)
 add_subdirectory(modeler)
 add_subdirectory(optimisation)
