+++ src/tests/src/solver/modelParser/CMakeLists.txt
@@ -11,6 +11,7 @@ add_executable(TestModelParser ${SOURCE_FILES})
 target_link_libraries(TestModelParser
         PRIVATE
         Boost::unit_test_framework
+        Antares::modelConverter
         Antares::modelParser
         Antares::antares-solver-libObjectModel
 )
